pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Fast auth backend - бекенд для аутентификации и авторизации."
authors = [
    {name = "Arduinum628",email = "message.chaos628@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "ruff (>=0.12.8,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
seed-fake-users = "app.scripts.seed_fake_users:start_seed_users"
